Use only test files for subsystem benchmarks
