+++ polkadot/node/subsystem-bench/src/subsystem-bench.rs
@@ -16,35 +16,29 @@
 
 //! A tool for running subsystem benchmark tests designed for development and
 //! CI regression testing.
+
+use crate::{approval::bench_approvals, core::display::display_configuration};
+use approval::ApprovalsOptions;
+use availability::{prepare_test, DataAvailabilityReadOptions, NetworkEmulation, TestState};
 use clap::Parser;
-
-use colored::Colorize;
-
+use clap_num::number_range;
 use color_eyre::eyre;
-use pyroscope::PyroscopeAgent;
-use pyroscope_pprofrs::{pprof_backend, PprofConfig};
-
-use std::path::Path;
-
-pub(crate) mod approval;
-pub(crate) mod availability;
-pub(crate) mod cli;
-pub(crate) mod core;
-mod valgrind;
-
-const LOG_TARGET: &str = "subsystem-bench";
-
-use availability::{prepare_test, NetworkEmulation, TestState};
-use cli::TestObjective;
-
+use colored::Colorize;
 use core::{
 	configuration::TestConfiguration,
 	environment::{TestEnvironment, GENESIS_HASH},
 };
+use pyroscope::PyroscopeAgent;
+use pyroscope_pprofrs::{pprof_backend, PprofConfig};
+use serde::{Deserialize, Serialize};
+use std::path::Path;
 
-use clap_num::number_range;
+pub(crate) mod approval;
+pub(crate) mod availability;
+pub(crate) mod core;
+mod valgrind;
 
-use crate::{approval::bench_approvals, core::display::display_configuration};
+const LOG_TARGET: &str = "subsystem-bench";
 
 fn le_100(s: &str) -> Result<usize, String> {
 	number_range(s, 0, 100)
@@ -60,9 +54,6 @@ struct BenchCli {
 	#[arg(long, value_enum, ignore_case = true, default_value_t = NetworkEmulation::Ideal)]
 	/// The type of network to be emulated
 	pub network: NetworkEmulation,
-
-	#[clap(flatten)]
-	pub standard_configuration: cli::StandardTestOptions,
 
 	#[clap(short, long)]
 	/// The bandwidth of emulated remote peers in KiB
@@ -100,42 +91,25 @@ struct BenchCli {
 	/// Enable Cache Misses Profiling with Valgrind. Linux only, Valgrind must be in the PATH
 	pub cache_misses: bool,
 
-	#[command(subcommand)]
-	pub objective: cli::TestObjective,
+	#[arg(required = true)]
+	/// Path to the test sequence configuration file
+	pub path: String,
+}
+
+/// Supported test objectives
+#[derive(Debug, Clone, clap::Parser, Serialize, Deserialize)]
+#[command(rename_all = "kebab-case")]
+pub enum TestObjective {
+	/// Benchmark availability recovery strategies.
+	DataAvailabilityRead(DataAvailabilityReadOptions),
+	/// Benchmark availability and bitfield distribution.
+	DataAvailabilityWrite,
+	/// Benchmark the approval-voting and approval-distribution subsystems.
+	ApprovalVoting(ApprovalsOptions),
+	Unimplemented,
 }
 
 impl BenchCli {
-	fn create_test_configuration(&self) -> TestConfiguration {
-		let configuration = &self.standard_configuration;
-
-		match self.network {
-			NetworkEmulation::Healthy => TestConfiguration::healthy_network(
-				self.objective.clone(),
-				configuration.num_blocks,
-				configuration.n_validators,
-				configuration.n_cores,
-				configuration.min_pov_size,
-				configuration.max_pov_size,
-			),
-			NetworkEmulation::Degraded => TestConfiguration::degraded_network(
-				self.objective.clone(),
-				configuration.num_blocks,
-				configuration.n_validators,
-				configuration.n_cores,
-				configuration.min_pov_size,
-				configuration.max_pov_size,
-			),
-			NetworkEmulation::Ideal => TestConfiguration::ideal_network(
-				self.objective.clone(),
-				configuration.num_blocks,
-				configuration.n_validators,
-				configuration.n_cores,
-				configuration.min_pov_size,
-				configuration.max_pov_size,
-			),
-		}
-	}
-
 	fn launch(self) -> eyre::Result<()> {
 		let is_valgrind_running = valgrind::is_valgrind_running();
 		if !is_valgrind_running && self.cache_misses {
@@ -152,98 +126,36 @@ impl BenchCli {
 			None
 		};
 
-		let mut test_config = match self.objective {
-			TestObjective::TestSequence(options) => {
-				let test_sequence =
-					core::configuration::TestSequence::new_from_file(Path::new(&options.path))
-						.expect("File exists")
-						.into_vec();
-				let num_steps = test_sequence.len();
-				gum::info!(
-					"{}",
-					format!("Sequence contains {} step(s)", num_steps).bright_purple()
-				);
-				for (index, test_config) in test_sequence.into_iter().enumerate() {
-					gum::info!(target: LOG_TARGET, "{}", format!("Step {}/{}", index + 1, num_steps).bright_purple(),);
-					display_configuration(&test_config);
+		let test_sequence = core::configuration::TestSequence::new_from_file(Path::new(&self.path))
+			.expect("File exists")
+			.into_vec();
+		let num_steps = test_sequence.len();
+		gum::info!("{}", format!("Sequence contains {} step(s)", num_steps).bright_purple());
+		for (index, test_config) in test_sequence.into_iter().enumerate() {
+			gum::info!(target: LOG_TARGET, "{}", format!("Step {}/{}", index + 1, num_steps).bright_purple(),);
+			display_configuration(&test_config);
 
-					match test_config.objective {
-						TestObjective::DataAvailabilityRead(ref _opts) => {
-							let mut state = TestState::new(&test_config);
-							let (mut env, _protocol_config) = prepare_test(test_config, &mut state);
-							env.runtime().block_on(availability::benchmark_availability_read(
-								&mut env, state,
-							));
-						},
-						TestObjective::ApprovalVoting(ref options) => {
-							let (mut env, state) =
-								approval::prepare_test(test_config.clone(), options.clone());
+			match test_config.objective {
+				TestObjective::DataAvailabilityRead(ref _opts) => {
+					let mut state = TestState::new(&test_config);
+					let (mut env, _protocol_config) = prepare_test(test_config, &mut state);
+					env.runtime()
+						.block_on(availability::benchmark_availability_read(&mut env, state));
+				},
+				TestObjective::ApprovalVoting(ref options) => {
+					let (mut env, state) =
+						approval::prepare_test(test_config.clone(), options.clone());
 
-							env.runtime().block_on(bench_approvals(&mut env, state));
-						},
-						TestObjective::DataAvailabilityWrite => {
-							let mut state = TestState::new(&test_config);
-							let (mut env, _protocol_config) = prepare_test(test_config, &mut state);
-							env.runtime().block_on(availability::benchmark_availability_write(
-								&mut env, state,
-							));
-						},
-						TestObjective::TestSequence(_) => todo!(),
-						TestObjective::Unimplemented => todo!(),
-					}
-				}
-				return Ok(())
-			},
-			TestObjective::DataAvailabilityRead(ref _options) => self.create_test_configuration(),
-			TestObjective::DataAvailabilityWrite => self.create_test_configuration(),
-			TestObjective::ApprovalVoting(_) => todo!(),
-			TestObjective::Unimplemented => todo!(),
-		};
-
-		let mut latency_config = test_config.latency.clone().unwrap_or_default();
-
-		if let Some(latency) = self.peer_mean_latency {
-			latency_config.mean_latency_ms = latency;
-		}
-
-		if let Some(std_dev) = self.peer_latency_std_dev {
-			latency_config.std_dev = std_dev;
-		}
-
-		// Write back the updated latency.
-		test_config.latency = Some(latency_config);
-
-		if let Some(connectivity) = self.connectivity {
-			test_config.connectivity = connectivity;
-		}
-
-		if let Some(bandwidth) = self.peer_bandwidth {
-			// CLI expects bw in KiB
-			test_config.peer_bandwidth = bandwidth * 1024;
-		}
-
-		if let Some(bandwidth) = self.bandwidth {
-			// CLI expects bw in KiB
-			test_config.bandwidth = bandwidth * 1024;
-		}
-
-		display_configuration(&test_config);
-
-		let mut state = TestState::new(&test_config);
-		let (mut env, _protocol_config) = prepare_test(test_config, &mut state);
-
-		match self.objective {
-			TestObjective::DataAvailabilityRead(_options) => {
-				env.runtime()
-					.block_on(availability::benchmark_availability_read(&mut env, state));
-			},
-			TestObjective::DataAvailabilityWrite => {
-				env.runtime()
-					.block_on(availability::benchmark_availability_write(&mut env, state));
-			},
-			TestObjective::TestSequence(_options) => {},
-			TestObjective::ApprovalVoting(_) => todo!(),
-			TestObjective::Unimplemented => todo!(),
+					env.runtime().block_on(bench_approvals(&mut env, state));
+				},
+				TestObjective::DataAvailabilityWrite => {
+					let mut state = TestState::new(&test_config);
+					let (mut env, _protocol_config) = prepare_test(test_config, &mut state);
+					env.runtime()
+						.block_on(availability::benchmark_availability_write(&mut env, state));
+				},
+				TestObjective::Unimplemented => todo!(),
+			}
 		}
 
 		if let Some(agent_running) = agent_running {
